+++ src/main/java/kz/trip/travelmanagement/repository/ReviewRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
-    List<Review> findByTourId(int tourId);
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+    List<Review> findByTourId(long tourId);
 }
