+++ src/main/java/kz/trip/travelmanagement/repository/RoleRepository.java
@@ -3,8 +3,9 @@ package kz.trip.travelmanagement.repository;
 import kz.trip.travelmanagement.models.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Locale;
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByName(String name);
 }
